June/services/june-tts/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-keycloak==3.7.0
PyJWT==2.8.0
httpx==0.25.2
python-magic==0.4.27
librosa==0.10.1
soundfile==0.12.1
